Added context to MXMLParser, so we can set the model as the context and have access to it in the init methods.  Added context: parameter to -initWithXMLAttributes:context: so that objects have access to the model.  (Needed for MMPoint, which needs to be able to set the equation.)
Added method to MModel to get an equation by name.  This assumes that
all equation names are unique, which I'll need to check.  (It
certainly isn't enforced by the app.)  Otherwise we'll need to store
the group as well in the XML file.

We can now unarchive points (but not slope-ratios) from a transition's
point-or-slopes element.