Keep the old FormulaExpression, FormulaTerminal around in minimal form so we can use it to unarchive the old typedstreams and generate a string for the equation.
Added new MMFormulaExpression and MMFormulaTerminal replacements, that
are a bit simpler than the original.  Changed these to make sure they
print parenthesis in the equation when needed.  They only print them
when needed, so the result looks a bit different than the old classes.

Removed unused cacheTag, cacheValue ivars.

We still need to change the old unarchiving so the old classes are
unarchived, generate the expression string, and then get reparsed into
the new classes.