Moved the OVERSAMPLING_OSCILLATOR define into the C file, since that is the only place that it is used.
Rearranged the structure members, grouping those that are only used
for the PULSE waveform.  Added a waveform member, so that
TRMWavetableUpdate() can check this to see if it should update (for
PULSE only).  That means that the calling code won't need to do the
check.

Moved the two local functions to the bottom of the file.  Changed type
of 'j' from int to double in one place.