Store the posture in Event when isAtPosture.
Since I don't always know the correct posture when I create the event, fill them
in after applying all of the rules.

Currently there is an assertion in place to make sure this event is the same as
what the previous code had.  Next up is removing the old code.