Protect against zero denominator when drawing smooth graph.
This happens when two intonation points have the same absolute time.

Draw a vertical red bar and log a warning to indicate there is a problem, and
don't crash at least.

We should protect against this happening on data entry, but there are several
ways the time can be altered.

EventList code needs to be checked for similar issue.